+++ mm/2s2h/BenGui/LocalMouseStateManager.h
