+++ mm/2s2h/Enhancements/Camera/FreeLook.cpp
@@ -2,7 +2,8 @@
 #include "2s2h/GameInteractor/GameInteractor.h"
 #include "2s2h/ShipInit.hpp"
 #include "CameraUtils.h"
-#include "2s2h/Enhancements/Controls/Mouse/Mouse.h"
+#include "2s2h/Mouse.h"
+#include "2s2h/Enhancements/Controls/Mouse/MouseShieldAim.h"
 
 extern "C" {
 #include "macros.h"
