+++ mm/2s2h/Enhancements/Controls/Mouse/MouseShieldAim.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "z64camera.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void HandleShieldCameraControl(Camera* camera, s16 viewYaw);
+
+#ifdef __cplusplus
+} // extern "C"
+#endif
